master-agent/agents/translator_agent/pyproject.toml
[project]
name = "genai-agents"
version = "0.1.0"
description = "Multi-language detection system (Variation 2)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "AgentMaker", email = "agents@agentmaker.dev"}
]
keywords = ["nlp", "language"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "genai-protocol",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "loguru>=0.7.3",
    "openai>=1.70.0",
    "pydantic==2.11.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0"
]

[project.urls]
Homepage = "https://agentmaker.dev"
Documentation = "https://docs.agentmaker.dev"
Repository = "https://github.com/agentmaker/agents"
Issues = "https://github.com/agentmaker/agents/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"
